pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentevals"
version = "0.4.1"
description = "Standalone framework to evaluate agent correctness based on portable OpenTelemetry traces"
requires-python = ">=3.11"
dependencies = [
    "google-adk[eval]>=1.25.0",
    "click>=8.0",
    "tabulate>=0.9.0",
    "pandas>=2.2.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "opentelemetry-proto>=1.36.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "agentevals-grader-sdk>=0.1.0",
]

[project.optional-dependencies]
live = [
    "mcp>=1.26.0",
    "httpx>=0.27.0",
]
streaming = [
    "opentelemetry-sdk>=1.20.0",
    "websockets>=12.0",
]

[project.scripts]
agentevals = "agentevals.cli:main"

[tool.hatch.build]
artifacts = ["src/agentevals/_static/**"]

[tool.hatch.build.targets.wheel]
packages = ["src/agentevals"]

[tool.uv]
no-build-isolation-package = ["rouge_score"]

[tool.uv.workspace]
members = ["packages/grader-sdk-py"]

[tool.uv.sources]
agentevals-grader-sdk = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: OTLP pipeline tests with ASGI apps (no API keys)",
    "e2e: End-to-end tests requiring API keys and real agents",
]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.11.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120
fix = true
src = ["src", "packages", "tests"]

[tool.ruff.lint.isort]
known-first-party = ["agentevals"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "Q", "I", "ASYNC", "T20"]
ignore = ["F401", "E501", "B008", "ASYNC109"]
